Refactor temperature schedules: Int args, composable zig_zag
- Rename params: passed_time→step, total_time→iterations
- Change first two args from Float64 to Int (iteration counts)
- zig_zag now takes cooling function as first arg (zig_cool)
- Add docstrings with reference link

+++ src/modules/Algorithms/src/simulated_annealing.jl
@@ -15,7 +15,7 @@ step size σ_t scaling as `σ_t_init * √(T/T_init)` (Langevin dynamics justifi
 
 # Example
 ```julia
-schedule = (step, total) -> quadratic_additive(Float64(step), Float64(total), 10.0, 0.01)
+schedule = (step, total) -> quadratic_additive(step, total, 10.0, 0.01)
 algorithm = SimulatedAnnealing(
     x -> my_energy(x),
     (x, σ) -> perturb_single_point(x, σ),
